Kody powoli jak żółw ociężale wpadją do gita po łączach ospale.

+++ src/SygnalRadiowy/IDIgitalSignal.java
@@ -4,7 +4,7 @@ package SygnalRadiowy;
  * Created by admin on 31.08.2017.
  */
 public interface IDIgitalSignal {
-    double[] getDigit();
+    int[] getDigit();
     void setDigit(int [] digitData);
     void printDigit();
 }
